Allow specifying of underlying rev in fixtures
The change in the previous commit revealed a bug in the test framework:
when we started passing the full paired version back from a lock, rather
than just the human-readable version, it caused that locked version to
not get excluded by the logic in versionQueue.advance(), as lock
versions had revs that manifest versions did not.

The change here allows specification of underlying revs on both sides.
It's a bit more verbose, but it helps keep us more honest.