Trigger prefetching when advantageous in solver
There are two, somewhat-overlapping goals here:

1. Don't initiate network activity, even in the background, unless we're
pretty sure we're going to need the results of that work.
2. Don't initiate background network activity unless we're pretty sure
it will result in beneficial parallelism.

Basically, we want to avoid network activity if we can. If we're likely
to need it, though, then to avoid complexity growth owing to doing a ton
of background work (which entails synchronization), we do async
prefetching only if we have a strong reason to believe it will result in
parallelism.