Move alternate net addr tests to bimodal
Basic tests make the implicit assumption that declared name == package
name (and also that each project is one and only one package); this is
one of the key ways that allows basic tests to operate while ignoring
the bimodal reality. This makes them poorly suited to the more dynamic
requirements of local/net-name swapping.