Rebalance test case for ordering purposes
So, this is gnarly.

This test case was erroneously passing because of the order in which
dependencies were being processed - with other factors held constant,
the transitive dep (with the alternate net name specification) was
actually being reached BEFORE the unconstrained/default one specified
closer to the root.

This just reinforces the TODO on the test - while it seems unlikely that
this situation will occur _right_ now, this is a really nasty gotcha
that we want to patch up ASAP. These kinds of ordering issues literally
mean that someone pushing a new version of a dep could cause your build
to fail, simply because that additional version being present causes the
solver to visit unselected identifiers in a different order.