Introduce gopkginSource, plus its maybe
gopkginSource are basically thin wrappers around a gitSource. The main
difference is that the gopkginSource performs filtering on the versions
that come through according to the constraint stated in the import path
URL.

This violates both name uniqueness and name/version orthogonality, but
there's not really a ton we can do - we have to at least somewhat
respect the guidelines that gopkg.in set up, as they represent valid
user intent.