Handle malformed go source case in ListPackages()
Prior to this, encountering malformed Go source code in any package
woudl cause the entire ListPackages operation to fail. Now, when the Go
source text scanner encounters an error, we catch it and store it
correctly in a PackageOrErr.

This may well not be exhaustive - we may need to cover other types.
Handles one aspect of #99.