Split out 'rootdata' struct from solver
This separates a bunch of the static state/rules/information that comes
from the root project and input parameters into a discrete subsystem.
The only real benefit here is focusing the state tracked by the solver
in on the actual algorithm of solving, and less so these static rules -
which should make it a bit easier for other people to grok.