Allow import cycles to avoid xtest import problems
Because xtests (package *_test) can import the package they cohabit a
directory with, but ExternalReach() merges normal and test imports
together (and ListPackages() merges test and xtest imports into test
imports), it's possible for it to appear like there's an import cycle,
even though there isn't one in code that would actually compile
together.

The proper solution to this requires refactoring to allow multiple
packages to exist per directory. That's a major undertaking, and really
should only be attempted as part of #99. So, this is a quick fix in the
meantime.