Fix bug in preparation of workmaps
The old logic was a holdover from before a proper depth-first search
algorithm was implemented in wmToReach(). It was trying to do a "bit" of
the search work before the real algorithm; however, the final else
statement was causing some internal imports to be dropped if the
referent had already been visited. The bug was intermittent, as it
depended on map iteration order.

This also solves the secondary problem of inaccurate
backpropagation/poisoning in wmToReach itself, as the internal linkage
data it's operating on is now reliable.