Re-enable multi fetch test
This is a great milestone. Reliable passage of that test demonstrates
that the sourceMgr can not only manage very heavy concurrency and
strongly suggest that the sourceGateway's dependent ordering logic is
correct (within the four included methods, anyway), but it also
that Release() correctly waits for all subprocess calls to end before
returning.