Make the vcs_repo tests fail when encountering an error
Calling `t.Error` is not enough when encountering an error. While this
marks the test as failed, it won't actually prevent the test from
continuing to run with invalid state. I've replaced these instances with
`t.Fatal`.

There looks to be plenty of opportunity to refactor these tests, but
that's better done in a follow up PR. I've made the smallest change
possible to make these tests not panic in case of errors.